requirements.txt
groq==0.9.0
python-dotenv==1.0.0
streamlit==1.40.2
spotipy==2.23.0
